backend/requirements.txt
api==0.0.7
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
Django==3.2
django-colorfield==0.7.2
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
idna==3.4
importlib-metadata==5.0.0
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
nose==1.3.7
oauthlib==3.2.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.5.0
python-slugify==6.1.2
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.12
zipp==3.8.1
gunicorn==20.1.0
python-dotenv==0.20.0
Pillow==9.2.0
psycopg2-binary==2.9.3
